fix(id): replace int with Long type for ids

+++ src/main/java/com/medilabo/abernathyclinic/patient/repository/StreetRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.medilabo.abernathyclinic.patient.entity.City;
 import com.medilabo.abernathyclinic.patient.entity.Street;
 
-public interface StreetRepository extends JpaRepository<Street, Integer> {
+public interface StreetRepository extends JpaRepository<Street, Long> {
 	
 	@Query("select s from Street s where s.name = :name and s.city = :city")
 	public Optional<Street> findByNameAndCity(String name, City city);
